% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_utils.R
\name{merge_euvar}
\alias{merge_euvar}
\title{Merge EU variables into a user dataset}
\usage{
merge_euvar(user_df, vars = NULL)
}
\arguments{
\item{user_df}{Data.frame containing columns \code{country.a}, \code{country.b}, \code{year}.}

\item{vars}{Vector of variable names from EU dataset to merge.}
}
\value{
Original data.frame with selected EU variables merged by dyad/year.
}
\description{
Merge EU variables into a user dataset
}
